# UserLogin/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "1.0.0",
    "title": "AI Quiz App API",
    "description": "API for managing questions and answers for AI-powered classroom quizzes",
    "contact": {
      "name": "API Support",
      "email": "mdoswell@my.bcit.ca"
    }
  },
  "host": "https://octopus-app-x9uen.ondigitalocean.app",
  "basePath": "/",
  "tags": [
    {
      "name": "Authentication",
      "description": "Login and authentication API"
    },
    {
      "name": "Quiz Session",
      "description": "API for managing quiz sessions/rooms where teachers can show questions and students can answer them"
    },
    {
      "name": "Question",
      "description": "API for managing quiz questions"
    },
    {
      "name": "Answer",
      "description": "API for managing quiz answer"
    },
    {
      "name": "AI Services",
      "description": "API for accessing AI services for transcribing question audio and grading answer correctness"
    },
    {
      "name": "API Usage",
      "description": "API for accessing statistics on API usage"
    }
  ],
  "schemes": [
    "http"
  ],
  "securityDefinitions": {},
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/createsession": {
      "post": {
        "description": "",
        "tags": ["Quiz Session"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "User JWT"
                  }
                },
                "required": ["token"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/joinsession": {
      "post": {
        "description": "",
        "tags": ["Quiz Session"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "User JWT"
                  },
                  "session_code": {
                    "type": "string",
                    "description": "Code of the quiz session to join"
                  }
                },
                "required": ["token", "session_code"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/retrivequestion": {
      "post": {
        "description": "Get the current question of the session that the user is currently in.",
        "tags": ["Question"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "User JWT"
                  },
                  "sessionId": {
                    "type": "string",
                    "description": "Code of the quiz session to join"
                  }
                },
                "required": ["token", "sessionId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/recieveanswer": {
      "post": {
        "description": "Send an answer for the current question of the current session.",
        "tags": ["Answer"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "User JWT"
                  },
                  "questionId": {
                    "type": "string",
                    "description": "Id of the question to answer"
                  },
                  "question": {
                    "type": "string",
                    "description": "Text of the question to answer"
                  },
                  "answer": {
                    "type": "string",
                    "description": "Text of the answer to the question"
                  }
                },
                "required": ["token", "question", "questionId", "answer"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/checksession": {
      "post": {
        "description": "Check if the given session is active.",
        "tags": ["Quiz Session"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "User JWT"
                  },
                  "sessionCode": {
                    "type": "string",
                    "description": "Join code of the session to check"
                  }
                },
                "required": ["token", "sessionCode"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/destroysession": {
      "put": {
        "description": "End the given session.",
        "tags": ["Quiz Session"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "User JWT"
                  },
                  "sessionId": {
                    "type": "string",
                    "description": "Id of the session to end"
                  }
                },
                "required": ["token", "sessionId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/confirmquestion": {
      "post": {
        "description": "Confirms the current question of the session that the user is in (adds the question to the database).",
        "tags": ["Question"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "User JWT"
                  },
                  "sessionId": {
                    "type": "string",
                    "description": "Id of the session to which question will be added"
                  },
                  "question": {
                    "type": "string",
                    "description": "Text of the question to answer"
                  }
                },
                "required": ["token", "sessionId", "question"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/endquestion": {
      "post": {
        "description": "End the answering period for the current question.",
        "tags": ["Quiz Session"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "User JWT"
                  },
                  "sessionId": {
                    "type": "string",
                    "description": "Id of the session for which question will be ended"
                  }
                },
                "required": ["token", "sessionId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/getanswers": {
      "post": {
        "description": "Get all answers to the current question of the given session.",
        "tags": ["Answer"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "User JWT"
                  },
                  "sessionId": {
                    "type": "string",
                    "description": "Id of the session for which answers will be retrieved"
                  }
                },
                "required": ["token", "sessionId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/signup": {
      "post": {
        "description": "Create an account for the specified user.",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of new user"
                  },
                  "email": {
                    "type": "string",
                    "description": "Email address of the new user"
                  },
                  "password": {
                    "type": "string",
                    "description": "Password of the new user"
                  }
                },
                "required": ["name", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/login": {
      "post": {
        "description": "Log the user into the app (creates a JWT for the user).",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "Email address of the user"
                  },
                  "password": {
                    "type": "string",
                    "description": "Password of the user"
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/logout": {
      "delete": {
        "description": "Logs the user out of the app (removes their token from the active session database).",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "User JWT"
                  }
                },
                "required": ["token"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/checktoken": {
      "post": {
        "description": "Check validity of a user JWT.",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "User JWT"
                  }
                },
                "required": ["token"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/transcribeQuestion": {
      "post": {
        "description": "Uses AI to transcribe the given audio file to question text.",
        "tags": ["AI Services"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/apiEndpointUsage": {
      "post": {
        "description": "Gets the total number of requests to each endpoint of the API.",
        "tags": ["API Usage"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/apiUserUsage": {
      "post": {
        "description": "Gets the total number of requests to the API for each user.",
        "tags": ["API Usage"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "errorResponse.400": {
      "type": "object",
      "properties": {
        "code": {
          "type": "number",
          "example": 400
        },
        "message": {
          "type": "string",
          "example": "The request was malformed or invalid. Please check the request parameters."
        }
      }
    },
    "errorResponse.401": {
      "type": "object",
      "properties": {
        "code": {
          "type": "number",
          "example": 401
        },
        "message": {
          "type": "string",
          "example": "Authentication failed or user lacks proper authorization."
        }
      }
    },
    "errorResponse.403": {
      "type": "object",
      "properties": {
        "code": {
          "type": "number",
          "example": 403
        },
        "message": {
          "type": "string",
          "example": "You do not have permission to access this resource."
        }
      }
    },
    "errorResponse.404": {
      "type": "object",
      "properties": {
        "code": {
          "type": "number",
          "example": 404
        },
        "message": {
          "type": "string",
          "example": "The requested resource could not be found on the server."
        }
      }
    },
    "errorResponse.500": {
      "type": "object",
      "properties": {
        "code": {
          "type": "number",
          "example": 500
        },
        "message": {
          "type": "string",
          "example": "An unexpected error occurred on the server. Please try again later."
        }
      }
    }
  }
}